[None][chore] cast guided draft gather index to int64
num_accepted_tokens is allocated int32 in
SpecWorkerBase.sample_and_accept_draft_tokens, so gidx is int32.
torch.gather accepts int32 indices on the pinned torch range (>=2.11)
but raises on older versions; cast to int64 defensively and for
consistency with the existing .long() used for the same gather earlier
in the worker (dspark.py:356).

+++ tensorrt_llm/_torch/speculative/dspark.py
@@ -426,7 +426,7 @@ class DSparkWorker(SpecWorkerBase):
         else:
             full_logits = gen_logits.transpose(0, 1).contiguous()
 
-        gidx = (num_accepted_tokens - 1).clamp(min=0).unsqueeze(1)
+        gidx = (num_accepted_tokens - 1).clamp(min=0).unsqueeze(1).long()
         new_tokens = accepted_tokens.gather(1, gidx).squeeze(1).to(torch.int32)
 
         gen_draft_tokens = []
